refactoring changes for construction task to make way for cooking task
